visora_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aif360==0.6.1
fairlearn==0.10.0
shap==0.45.1
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
google-genai==1.0.0
reportlab==4.2.2
langgraph==0.2.28
langchain-core==0.3.15
python-dotenv==1.0.1
websockets==13.1
